refactor(element quiz): rename the folder periodic to element
