+++ openday_scavenger/puzzles/element/services.py
