+++ openday_scavenger/puzzles/element/static/css/styles.css
