+++ openday_scavenger/puzzles/element/static/data/element_list.json
