+++ openday_scavenger/puzzles/element/static/js/index.js
