+++ openday_scavenger/puzzles/element/static/sql/initiate.sql
